--- modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/CrcCheckingDataInput.java
@@ -1,2 +1,0 @@
-package org.apache.ignite.internal.processors.cache.persistence.wal.io;public class CrcCheckingDataInput {
-}
