+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/serializer/RecordV2Serializer.java
@@ -30,7 +30,6 @@ import org.apache.ignite.internal.processors.cache.persistence.wal.ByteBufferBac
 import org.apache.ignite.internal.processors.cache.persistence.wal.SegmentEofException;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WALPointer;
 import org.apache.ignite.internal.processors.cache.persistence.wal.WalSegmentTailReachedException;
-import org.apache.ignite.internal.processors.cache.persistence.wal.io.FileInput;
 import org.apache.ignite.internal.processors.cache.persistence.wal.serializer.io.RecordIO;
 import org.apache.ignite.internal.util.GridUnsafe;
 import org.apache.ignite.internal.util.typedef.F;
@@ -246,10 +245,7 @@ public class RecordV2Serializer implements RecordSerializer {
 
     /** {@inheritDoc} */
     @Override public WALRecord readRecord(ByteBufferBackedDataInput in, WALPointer expPtr) throws IOException, IgniteCheckedException {
-        if (in instanceof FileInput)
-            return RecordV1Serializer.readWithCrc((FileInput)in, expPtr, recordIO);
-        else
-            return RecordV1Serializer.readWithoutCrcCheck(in, expPtr, recordIO);
+        return RecordV1Serializer.readWithCrc(in, expPtr, recordIO);
     }
 
     /**
